+++ tests/test_model_vis.py
@@ -6,9 +6,8 @@ from opstool.preprocessing import gen_grav_load
 from opstool.vis import GetFEMdata, OpsVisPlotly, OpsVisPyvista
 from opstool import load_ops_examples
 
-
-load_ops_examples("ArchBridge")
-# load_ops_examples("CableStayedBridge")
+# load_ops_examples("ArchBridge")
+load_ops_examples("CableStayedBridge")
 # load_ops_examples("Dam")
 # load_ops_examples("Frame3D")
 # load_ops_examples("Igloo")
@@ -19,26 +18,26 @@ ModelData = GetFEMdata()
 ModelData.get_model_data()
 ModelData.get_eigen_data(mode_tag=15)
 opsv = OpsVisPlotly(point_size=2, line_width=2, colors_dict=None, theme="plotly",
-                    color_map="jet", on_notebook=False, results_dir="opstool_output")
-opsv.model_vis(show_node_label=True, show_ele_label=True,
-               show_local_crd=False, label_size=8,
+                     color_map="jet", on_notebook=False, results_dir="opstool_output")
+opsv.model_vis(show_node_label=False, show_ele_label=False,
+               show_local_crd=True, label_size=8,
                show_outline=True,
-               opacity=1.0,)
-opsv.eigen_vis(mode_tags=[2, 15], subplots=False,
-               alpha=None, show_outline=True,
-               show_origin=True, opacity=1.0,
+               opacity=1.0, )
+opsv.eigen_vis(mode_tags=[1, 15], subplots=False,
+               alpha=None, show_outline=False,
+               show_origin=False, opacity=1.0,
                show_face_line=True)
 
 # opsv.eigen_anim(mode_tag=4, alpha=None, show_outline=False,
 #                 opacity=1, framerate=3,
 #                 show_face_line=True,
-#                 save_file="EigenAnimation.gif")
+#                 save_html="EigenAnimation")
 
 # responses
 gen_grav_load(ts_tag=1, pattern_tag=1,
               factor=-9.81, direction="Z")
 # 分析参数设置
-Nsteps = 10   # 分析10步
+Nsteps = 10  # 分析10步
 ops.wipeAnalysis()
 ops.system('BandGeneral')
 ops.constraints('Transformation')
@@ -51,24 +50,22 @@ ops.analysis('Static')
 ModelData.reset_steps_state()
 for i in range(Nsteps):
     a = ops.analyze(1)
-    ModelData.get_node_resp_step(analysis_tag=1,
-                                 num_steps=Nsteps,
+    ModelData.get_node_resp_step(num_steps=Nsteps,
                                  model_update=False)
-    ModelData.get_frame_resp_step(analysis_tag=1, num_steps=Nsteps,)
-opsv.deform_vis(analysis_tag=1, slider=True,
+    ModelData.get_frame_resp_step(num_steps=Nsteps)
+opsv.deform_vis(slider=True,
                 response="disp", alpha=None,
                 show_outline=False, show_origin=False,
                 show_face_line=False, opacity=1,
                 model_update=False)
-# opsv.deform_anim(analysis_tag=1,
-#                  save_fig="yan.gif",
-#                  response="accel", alpha=None,
-#                  show_outline=False,
-#                  show_face_line=False, opacity=1,
-#                  framerate=3,
-#                  model_update=False)
-opsv.frame_resp_vis(analysis_tag=1,
-                    ele_tags=None,
+opsv.deform_anim(
+                 # save_fig="yan.gif",
+                 response="disp", alpha=None,
+                 show_outline=False,
+                 show_face_line=False, opacity=1,
+                 framerate=3,
+                 model_update=False)
+opsv.frame_resp_vis(ele_tags=None,
                     slider=True,
                     response="my",
                     show_values=False,